requirements.txt
requests
beautifulsoup4
spotipy
python-dotenv
pillow
ytmusicapi
PySide6
rapidfuzz